removed prints and warnings
